requirements.txt
paho-mqtt
pandas
matplotlib
numpy